Add MenuStrip tests to UI test project
